Add NCG validation experiments with CSNN/SoftHebb features
- Add NCG (Neuron-Centric Growth) dynamic neuron allocation experiments
  on merged CIFAR10 classes (5-class) and standard CIFAR10 (10-class)
- Feature extractors: STDP-CSNN and SoftHebb CNN
- Compare NCG vs static neuron allocation baselines (static-50/60/100/200)
- Fix SoftHebb run_cifar.py for PyTorch 2.8 compatibility
  (LRScheduler.step() API change, cuda→cpu fallback)
- Add .gitignore entries for large data files (npy, extracted features,
  raw datasets, tensorboard events, venv)
- Experiment scripts: prepare_dataset_*.py, compare_experiments*.py
- Results: NCG achieves best accuracy with fewer total neurons by
  selectively allocating more neurons to harder classes

+++ ft-extract/STDP-CSNN/csnn/include/ClassParameter.h
@@ -386,7 +386,7 @@ public:
 		if(_initialized) {
 			throw std::runtime_error("Variable already initialized");
 		}
-		_shape = Shape({is...});
+		_shape = Shape({static_cast<size_t>(is)...});
 	}
 
 	virtual void load(std::istream& stream) {
